refactor: moved api-logic to another directory
