+++ src/graphql/mutations/codeforces_status.rs
@@ -1,4 +1,4 @@
-use super::leaderboard_api::fetch_and_update_codeforces_stats;
+use crate::graphql::api::leaderboard_api::fetch_and_update_codeforces_stats;
 use async_graphql::{Context, Object, Result};
 use sqlx::PgPool;
 use std::sync::Arc;
