+++ src/graphql/mutations/leetcode_status.rs
@@ -1,4 +1,4 @@
-use super::leaderboard_api::fetch_and_update_leetcode;
+use crate::graphql::api::leaderboard_api::fetch_and_update_leetcode;
 use async_graphql::{Context, Object, Result};
 use sqlx::PgPool;
 use std::sync::Arc;
