+++ src/graphql/mutations/mod.rs
@@ -1,6 +1,5 @@
 pub mod attendance_mutations;
 pub mod codeforces_status;
-pub mod leaderboard_api;
 pub mod leetcode_status;
 pub mod member_mutations;
 pub mod project_mutations;
@@ -8,9 +7,6 @@ pub mod streak_mutations;
 
 pub use attendance_mutations::AttendanceMutations;
 pub use codeforces_status::FetchCodeForces;
-pub use leaderboard_api::fetch_and_update_codeforces_stats;
-pub use leaderboard_api::fetch_and_update_leetcode;
-pub use leaderboard_api::update_leaderboard_scores;
 pub use leetcode_status::FetchLeetCode;
 pub use member_mutations::MemberMutations;
 pub use project_mutations::ProjectMutations;
